Remove deprecated errors for new NavigationErrors
* Move internal navigation implementations to their own file
