+++ SitumWayfinding/Classes/WayfindingLibrary/Protocols/OnNavigationListener.swift
@@ -184,32 +184,3 @@ extension NavigationError: LocalizedError {
         }
     }
 }
-
-internal struct WYFNavigation: Navigation {
-    var status: NavigationStatus
-    var destination: Destination
-}
-
-internal struct WYFDestination: Destination {
-    var category: DestinationCategory
-    var point: SITPoint {
-        switch category {
-        case .poi(let poi):
-            return poi.position()
-        case .location(let point):
-            return point
-        }
-    }
-    var identifier: String?  {
-        guard case .poi(let poi) = category else { return nil }
-        return poi.identifier
-    }
-    var name: String? {
-        guard case .poi(let poi) = category else { return nil }
-        return poi.name
-    }
-
-    init(category: DestinationCategory) {
-        self.category = category
-    }
-}
